+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- `access` model added
+- `access` property added to `site`
+- `access` property added to `protocol`
+
 ### Changed
 
 ### Fixed